trim overspecified digits from durability calcs, add some other exp factors (#30)
* trim overspecified digits from durability calcs

However these were originally calculated, it looks like it wasn't done
at the precision which is implied by the given number of digits. I've
verified by doing the calculations with both mpmath and Wolfram Alpha.

To check my work, enter `GammaRegularized[n-k+1, n*0.1]` for the various
values of `n` and `k` on Wolfram Alpha.

Given that we are only given p to one significant digit (0.1) and is
clearly just a very loose estimate, most of these digits aren't actually
meaningful anyway, but they do help to illustrate the effect on
durability when n and k are changed.

* add some durability numbers at other exp factors

assuming these are the sorts of parameters stefan wanted.